Some refactoring making ability not a data object and testing the combat system.
